requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
environs==11.0.0
filelock==3.16.1
marshmallow==3.23.1
packaging==24.1
platformdirs==4.3.6
python-dotenv==1.0.1
sqlparse==0.5.1
virtualenv==20.27.1
